persist: add lease type to cursor, add into_lease
This allows extracting the lease back out and doing any necessary
cleanup. This will be required for the "free-standing" batch writer/read
API, where we have a method that can read a bunch of batches and the
batches themselves will be the lease. Those batches have to be deleted
after reading them.