storage: scope source/sink statistics to a replica
Before, with sources/sink running on multi-replica clusters (which are
not enabled in prod right now), we would overcount some metrics.
Specifically, there a gauge-metrics which would work correctly and there
are counter metrics where we would overcount. For example, each instance
of the source running on each replica would report `messages_received`.

Now, we instead report statistics per replica, that is we add a new
`replica_id` column to the statistics collection.

To achieve the above, we include a couple of refactorings. See below.

Before, we were using `*StatisticsUpdate` both to collect and transmit
statistics from cluster to the controller _and_ to keep those statistics
in the controller. Now we add a separate `Controller*Statistics`, which
has exactly the functionality needed in the controller, which now goes
beyond what the `Source/SinkStatisticsUpdate` supported before.

Also before, a bunch of features like initializing statistics with a
"zero row" where hard to understand in the code and so I added explicit
trait implementations and the code in the scraper should now be more
obvious to understand.

Closes https://github.com/MaterializeInc/database-issues/issues/9010
Closes https://github.com/MaterializeInc/database-issues/issues/9015