storage: delineate ControllerSourceStatistics from SourceStatisticsUpdate
Before, we were using `SourceStatisticsUpdate` both to collect and
transmit statistics from cluster to the controller _and_ to keep those
statistics in the controller. We want to make changes to how the
controller incorporates statistics updates from the clusters, so we need
to now separate these.

IMO, this in itself is already an improvement because
`SourceStatisticsUpdate` has a lot of functionality that is not needed
in the controller slash can be confusing when read in the context of the
controller.