storage: correct source statistics for multi-replica clusters
Before, with sources running on multi-replica clusters (which are not
enabled in prod right now), we would overcount some metrics.
Specifically, there a gauge-metrics which would work correctly and there
are counter metrics where we would overcount. For example, each instance
of the source running on each replica would report `messages_received`.

Now, we treat counter metrics differently: for those metrics where all
sources are reporting basically the same counter update we only take the
updates from the "first" replica. For gauges we just use all of them,
and the assumption is that these are largely the same on all replicas.

The one counter metric that is treated different from the others is
`updates_committed`: only one replica will manage to commit updates, so
we have to make sure we take this counter in from all replicas, to
capture all messages that have been committed.

Implements https://github.com/MaterializeInc/database-issues/issues/9010