k8s: don't unwrap when getting disk metrics
The new code for collecting disk metrics from orchestrated processes is
too liberal in its use of unwraps. Specifically, it is possible for the
get-service call to fail, in which case the unwrap on that call would
panic environmentd.

This is fixed by returning an error instead. The `get_disk_metrics`
function contained two additional unwraps, which are replaced as well.
They might be fine, but better to be safe than sorry.