storage/pg: switch to a per-row round-robin to ensure even distribution
We recently observed that distributing data container-by-container doesn't work
well if the size of containers isn't uniform and can lead to OOMs if the
distributing worker always happens to distribute a small container to itself
and a large container to everyone else.

We don't yet know why the oscillation occurs but until then this PR should help
with even distribution of work regardless of the size patterns.

I have also combined the exchange of data with the casting of rows in a single
operator.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>