storage/reclock: only reclock when there is a probe
Remove the logic where a bidning for the empty frontier is created without a
probe. This interacts badly with source operators that exit and advance their
output to the empty forniter.

One situation this can happen is if a source operator gets instantiated with no
subsources. Another way this can happen is if the source operator ingests a
table but then encounters an irrecoverable error for that table. In both
situations we would like the source operator to exit and lets its output
frontier advance to the empty frontier.

If in either of those situations we mint a reclock binding for the empty
frontier we lock ourselves up from ingesting more tables in the future. If the
source is altered to contain a subsource we will forever wait in the reclock
operator until all the data until the empty frontier is produced before
committing any data, which never happens.

The fix is to not mint any bindings when there is no probe. This fixes the
problem for the RLU strategy but the risk still exists for the old RIU
strategy. A fix for RIU is less obvious.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>