halt! when a lease expires in an acceptable way
We currently panic in a case that's known to be possible when a lease
expires. (A batch where the since is past the current since of a listen,
when that listen's lease is expired.)

We now check whether the lease has expired, and if so, trigger a less
noisy restart. Ideally, we'd percolate the error up so that we could do
something like restart a dataflow when that's possible... but we've got
a lot of callers here, and the particular case that we're running into
on the txns shard would need to halt anyways.