storage/pg: attempt exact snapshot count only when estimate is low enough
We have recently seen cases where the exact count of rows is problematic when
the target table contains many rows. The value of this exact count is not worth
the resources spent since the result only informs the snapshot progress reporting.

This PR changes the logic to always collect the estimated number of rows first
and if it's below a threshold (currently one million) attempt an exact count as
well. Otherwise we just use the estimate.

I retained the flag to compeltely disable exact counting in case the current
threshold is problematic.

This PR also removes the assertion that the total number of ingested rows match
the number of counted rows. This assertion effectively checks that PostgreSQL
transactions work, which doesn't seem very useful and due to other reasons
cause CI flakes. See MaterializeInc/database-issues#9362

Fixes MaterializeInc/database-issues#9362

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>