process: pull some setup out of `supervise_service_process`
This commit moves code around in the process orchestrator. Concretely,
it pulls some pieces out of the `supervise_service_process` task into
`ensure_service`:

 * collecting listen addresses
 * interpolating the command wrapper
 * constructing the service args

This simplifies the code a bit, but the main motivation is that it
enables passing also the peer listen addresses to the `args`
constructor, which is done in a subsequent commit.