process: drop existing processes on config change
This commit changes `ensure_service` to drop existing service processes
when a service config change is detected. Previously we would attempt to
reuse them and just create the missing ones. This doesn't work in a
world where each process needs to start up with a compatible Timely
configuration containing the addresses of all its peers.