test: retry statement log queries
The statement logging tests could fail because statement logging happens
async and they might not be waiting long enough for the log to fill.
This commit bumps the total wait time and also adds checks every second
so the tests don't wait unnecessarily.

The flaky tests previously had sleeps of 5s, while the test with a 10s
sleep seemed to not exhibit flakes, suggesting that 10s is enough of a
wait time.