test: unflake frontier-rounding.td
This commit unflakes frontier-rounding.td by disabling the tests that
check for frontier rounding _not_ working correctly. These tests were
meant to document the status quo but they aren't that useful, especially
when they are flaky.

As to why frontier-rounding.td was flaky before, I didn't manage to
reproduce it but it can randomly happen for a collection with
non-rounded frontiers to have a frontier that looks like it has been
rounded.