test: unflake test-lgalloc-limiter cluster test
In the test that ensures that the lgalloc limiter correctly kills the
replica when the disk limit is exceeded it could happen that the call to
`c.up("clusterd1")` would hang forever because the replica was killed
before it could report successful booting. Fixed by not making this call
wait for successful booting.