adapter: in support_execute_one drop SessionClient only after receiving result rows
As the comment in code explains, we would previously immediately cancel
the query before we have a chance to run the peek or read the results.

This went unnoticed because this code path is only used in our
telemetry/segment reporting, which we only do in production. I'm adding
a follow-up commit that enables at least some form of telemetry
reporting in testing.