cluster: don't panic on `LocalClient` send failure
The `LocalClient` made the incorrect assumption that its receiver would
never drop first. It is incorrect at least if the receiver is a compute
worker. Compute workers can learn about disconnects from other workers,
before their own GRPC server learns about the disconnect, in which case
they'll drop their receiver before the `LocalClient`'s sender is
dropped.

The fix is to not panic in the `LocalClient` on send failure but to
return an error instead. The error will make the GRPC server end the
connection and wait for a new one.