Improve error reporting for explain analyze (#32916)
The previous error reporting would report the index of the token as the
offset in the statement, which is incorrect. Fix this, and take the
opportunity to simplify the logic by using existing functions to
generate errors. This improves readability of the code slightly, but
doesn't produce necessarily better error messages. Seems OK because our
error messages aren't great anyway.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>