as-of-selection: fix pruning of dropped collections
The previously introduced work around for 0dt upgrades that prunes
dropped collections before performing as-of selection had a bug that
made it a no-op: It based the pruning on `self.collections` entries that
are dropped storage collections. However, `self.collections` only
contains compute collections, not storage collections, so the set of
pruned collections would always come out empty.

What we need to do instead is look for `self.collections` entries that
have dropped _storage inputs_ in the first pruning step. This commit
makes that change and also adds a test.