Compact command history when client connects (#32959)
Fix a corner case when connecting compute clients while updating
configuration meanwhile.

This change addresses a problem where we would send the command history
to a replica at the time of creating the instance, and not the history
at the time the replica connects. In tests, this caused surprising
behavior when adjusting flags: the replica would only pick up flag
changes after it restarted, which in some cases required multiple
restarts. This change seems beneficial in general because it allows us
to reduce the amount of commands provided to the replica before we
transmit it over the network.

The implementation is a minimal patch, but we could imagine as
follow-up:
* Instead of sharing a command channel between the controller and
instance task, share a command history.
* Add a pending state to the life cycle of an instance, and only
populate the initial commands once connected. This would have the added
benefit that we'd set the read holds closer the upper frontier that we
do now.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>