Limit overall memory utilization (#32939)
Limit the memory utilization of a cluster replica process, based on RSS
and swap utilization.

This change introduces a memory limiter task that periodically reads
`/proc/self/status` and extracts the `VmRSS` and `VmSwap` attributes. It
enforces that the sum of the two attributes is less than a configured
limit, taking into account an optional burst budget. It has the same
logic as the lgalloc limiter, off which it is based.

The first commit just copies the lgalloc limiter. The second changes it
and includes the scaffolding to integrate the memory limiter into the
cluster replica.

Fixes https://github.com/MaterializeInc/database-issues/issues/9436

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>