Temporal buckets (#31406)
Because, who doesn't like buckets?

Introduces `BucketChain`, a data structure to split time into log-sized
buckets.

A bucket chain presents as a mapping from time to storage. Over time,
the mapping gets more granular, with data that is closest to the last
extracted time is most granular. Its interface allows to find a storage
object for a given time, peel off all data until a frontier, and perform
internal maintenance.

A chain is well-formed if all buckets are within a factor of 4,
non-overlapping and without gaps. The chain does not need to be
well-formed for finding and peeling to be correct. Peeling might need to
do more work on the spot if the chain is not well-formed.

Restoring a chain to be well-formed can be an expensive operation as it
needs to split buckets. Splitting an individual bucket is
non-incremental, but we allow to supply fuel to incrementally restore
the overall chain property.

A simple benchmark shows its benefits: We maintain a list of 10 million
timestamps (now to now + 10_000_000s), and peel off the first 1000 in a
loop, until we read all elements. For a bucket chain, this takes around
440ms, and for vectors >24s on my computer. The PR encodes the benchmark
as a test.

## Notes

The current implementation only applies temporal bucketing after sources
in a dataflow. This isn't the only place where it might be useful, as
many MFPs could include temporal filters. We should cover all places,
but it's a bit more work to determine which places would actually
benefit. It is those that eventually lead to an arrangement, but
obtaining this information at the moment isn't trivial.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>