Temporal bucket: pass through data close to frontier
Adds a threshold parameter that influences which data we enter in the
bucket chain, and which data we pass through verbatim. It is a path summary
that we apply on the input frontier, and we only retain data
that has a timestamp not less or equal to the advanced frontier.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>