mz_join_v2: linear scan through times
This commit changes `mz_join_core_v2` to use DD's strategy for
generating join matches. For small inputs, a simple strategy is used
that is quadratic in the number of distinct times, whereas for large
inputs updates are sorted by time, so we can perform a linear scan over
time.

The helper data structures, `EditList` and `ValueHistory`, are lifted
from DD, but simplified a bit.