storage/upsert: remove the autospill backend
This commit rips out the autospill backend, which would start of using
the in-memory hashmap and then start spilling to RocksDB once some size
threshold was reached. We determined that this is an unnecessary
optimization and we can avoid some complexity by deleting it.

This also paves the way for replacing the in-memory hashmap with
RocksDB+MemEnv. Previously doing so would have been kind of weird
because the autospill backend would start "spilling" to an in-memory
RocksDB instance once the threshold is reached.