Temporal bucket: pass through data close to frontier (#33018)
Updates the temporal bucketing logic to only store data in the bucket
chain that is likely not to be revealed soon, as far as it's possible to
determine this locally. The condition we implement determines a upper
frontier based on the as-of of the dataflow, the current input frontier,
and a summary. It computes the join of the input frontier and the as-of,
and applies the summary to the result. The summary is configurable and
defaults to the equivalent of two seconds.

This results in a frontier that is always ahead of the input frontier,
causing the operator to release and pass through a configurable amount
of future updates, saving the cost associated with inserting and
extracting the data from the bucket chain.

This addresses the performance concerns for dataflows without future
retractions: most of the data will never enter the bucket chain. The
initial snapshot is within the 2s summary of the as-of, so it'll be
forwarded verbatim. Only if there are future updates, we'll store them
in the bucket chain.

There is a new cost associated with inserting data into the bucket
chain, but we hope to amortize this cost with the CPU savings associated
with having fewer outstanding updates in the merge batcher. We reduce
the cost from once every frontier tick to once for inserting and
extracting the data in the bucket chain, plus the maintenance of data in
the chain.

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>
Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>