Arrangement size logging: observe more batches
The arrangement size logging only records batches that are in the trace,
but not the ones that have been sent down the stream and that someone might
still hold on to. This change corrects this.

Note that we do not account for the size of arrangements that we currently
merge, we only account for the inputs but not the output. Differential
doesn't offer an API to observe the state of the newly formed batch.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>