adapter: remove unneeded replica options
Now that the Timely runtime parameters are passed as command line flags
to the clusterd processes, these SQL options passed when creating
unorchestrated replicas are dead code*:

* `STORAGE ADDRESSES`
* `COMPUTE ADDRESSES`
* `WORKERS`

This commit removes them from the adapter. It does not yet remove the
SQL parser support, to avoid breaking tests that want to run the same
replica DDL against different versions of Materialize.

*: Well, almost: The `COPY TO` sequencing code would like to know the
number of workers to decide, how many partitions to create. It's
probably not worth to keep the replica option just for that, given that
unorchestrated replicas are a test-only feature. So we just default to
the number of processes in this case.