SQL server: consistent snapshot and lsn selection (#32979)
The current mechanism for snapshot and LSN selection is not correct due
to the asynchronous nature of MS SQL CDC and the implementation details
of `SNAPSHOT` isolation.

This implementation is similar to the MySQL implementation in that it
locks the tables for a brief period to establish a point in time where a
consistent view of the DB can be observed. Once the `LSN` and
transaction sequence number (`XSN`), the table locks are dropped, and we
have a consistent version of the table(s) to read from as well as a
correct `LSN` to process CDC events.

### Motivation

Fixes https://github.com/MaterializeInc/database-issues/issues/9476

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.

---------

Co-authored-by: Petros Angelatos <petrosagg@gmail.com>