envd/ws: fix reception of initial websocket message
This commit fixes a bug in the `init_ws` logic where we would wait for
an initial authentication message in the loop, skipping over ping
messages, but forgetting to actually read a new message on each loop
iteration. As a result, if the first received message was a ping
message, we would busy-loop forever, potentially stalling out the entire
process by blocking the tokio runtime.