cluster: don't panic when dropping `TimelyContainer`
The panic was originally added to alert us to concurrency bugs where we
would accidentally drop the Timely runtime while trying to switch
between controller connections. The surrounding code has been simplified
a lot since, so the panic isn't as important as a safeguard anymore.

More importantly though, since the Timely runtime is now initialized
before we start serving controller connections, it is now actually
possible to drop the `TimelyContainer` without any bugs present:

 * Timely runtime is initialized, yielding a `TimelyContainer`.
 * The `TimelyContainer` is passed to `GrpcServer::serve`, which starts
   to listen for controller connections.
 * The listen fails, for example because the listen port is already in
   use.
 * `GrpcServer::serve` returns with an error, dropping the
   `TimelyContainer`.

In this scenario, the process exits anyway, so dropping the
`TimelyContainer` is fine and we shouldn't panic. The easiest fix is to
simply remove the panic.