misc: don't specify a memory limit for replicas by default
We are specifying a default memory limit of 4GiB for all replicas
created in mzcompose tests, even for replica sizes that don't explicitly
mention a limit (like '1'). This turns out to be a bit of a footgun as
tests might not expect the limit and become unreliable and/or slow by
getting their replicas intermittently killed.

The default memory limit was introduced to ensure that
`bin/environmentd` can be run without an unlimited license key. We can
also achieve that by specifically adjusting the replica sizes passed
under that command to have a default memory limit, while leaving the
sizes used by mzcompose unlimited.