sql: propagate retention windows by persisting them in SQL
We previously propagated source retention windows from the source to all the
subsources, forcing them to be the same. In the transactional DDL and "CREATE
TABLE FROM SOURCE" world this doesn't make sense. Each table from source is its
own object with its own retention window.

After this commit we always use the retention window used in the `CREATE TABLE
FROM SOURCE` statement and for the old syntax we propagate it during planning.
The end result is that each object contains its own retention window in the
catalog.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>